Add files via upload
Watermark:
In the context of data science or machine learning, a watermark typically refers to a technique used to protect the ownership or integrity of a dataset or model. watermarks serve as a form of digital rights management, helping to protect intellectual property and maintain trust and accountability in data-driven applications.

Watermarking Image Application

This project demonstrates how to add watermarks to images using Python and OpenCV. It includes two methods for adding watermarks:

Transparent Logo Method: This method adds a logo with transparency to an image using bitwise operations and ROI manipulation.
Semi-transparent Logo Method: This method adds a semi-transparent logo to an image using blending techniques.
Each method is thoroughly explained with step-by-step instructions and visualizations. The project serves as a useful resource for understanding image processing techniques and implementing watermarks programmatically.